feat(unsupported): plugin toolbar
