+++ src/serlo-editor/plugins/unsupported/index.tsx
@@ -1,7 +1,4 @@
-import { faWarning } from '@fortawesome/free-solid-svg-icons'
-
-import { FaIcon } from '@/components/fa-icon'
-import { useEditorStrings } from '@/contexts/logged-in-data-context'
+import { UnsupportedEditor } from './editor'
 import {
   EditorPlugin,
   EditorPluginProps,
@@ -9,7 +6,6 @@ import {
   scalar,
   string,
 } from '@/serlo-editor/plugin'
-import { selectSerializedDocument, store } from '@/serlo-editor/store'
 
 const unsupportedState = object({
   plugin: string(),
@@ -17,28 +13,10 @@ const unsupportedState = object({
 })
 
 export const unsupportedPlugin: EditorPlugin<typeof unsupportedState> = {
-  Component: UnsupportedRenderer,
+  Component: UnsupportedEditor,
   state: unsupportedState,
   config: {},
 }
 
 export type UnsupportedPluginState = typeof unsupportedState
-
-function UnsupportedRenderer(props: EditorPluginProps<UnsupportedPluginState>) {
-  const unsupportedStrings = useEditorStrings().plugins.unsupported
-
-  const unsupportedType = selectSerializedDocument(
-    store.getState(),
-    props.id
-  )?.plugin
-
-  return (
-    <div className="my-8 rounded-2xl bg-gray-100 p-4">
-      <b>
-        <FaIcon icon={faWarning} /> {unsupportedStrings.notSupported} [
-        {unsupportedType}]
-      </b>
-      <p className="mb-1 mt-3">{unsupportedStrings.explanation}</p>
-    </div>
-  )
-}
+export type UnsupportedPluginProps = EditorPluginProps<UnsupportedPluginState>
